# services/api/docs/swagger.yaml
components:
  schemas:
...
    auth.CredentialsAddBeginRequest:
      properties:
        device_name:
          type: string
      type: object
    auth.CredentialsAddCompleteRequest:
      properties:
        credential:
          type: object
        session_key:
          type: string
      type: object
...
    auth.OnboardBeginRequest:
      properties:
        device_name:
          type: string
        invite_id:
          type: string
        token:
          type: string
      type: object
    auth.OnboardBeginResponse:
      properties:
        publicKey:
          type: object
        session_key:
          type: string
        user_id:
          type: string
      type: object
    auth.OnboardCompleteRequest:
      properties:
        credential:
          type: object
        invite_id:
          type: string
        session_key:
          type: string
      type: object
    auth.PublicKeyOptionsResponse:
      properties:
        publicKey:
          type: object
        session_key:
          type: string
      type: object
...
info:
  title: ""
  version: ""
openapi: 3.1.0
paths:
...
  /api/v1/auth/credentials/add/begin:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.CredentialsAddBeginRequest'
        description: Begin credential add
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.PublicKeyOptionsResponse'
          description: OK
      summary: Add credential (begin)
      tags:
      - auth
  /api/v1/auth/credentials/add/complete:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.CredentialsAddCompleteRequest'
        description: Complete credential add
        required: true
      responses:
        "204":
          description: No Content
      summary: Add credential (complete)
      tags:
      - auth
...
  /api/v1/auth/onboard/begin:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.OnboardBeginRequest'
        description: Onboard begin request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.OnboardBeginResponse'
          description: OK
      summary: Onboard begin
      tags:
      - auth
  /api/v1/auth/onboard/complete:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.OnboardCompleteRequest'
        description: Onboard complete request
        required: true
      responses:
        "204":
          description: No Content
      summary: Onboard complete
      tags:
      - auth
...